fix make_zlib for mingw64 add test for all user_specific configurations standardization

+++ thirdParties/zlib/make_zlib.bat
@@ -10,9 +10,22 @@
 %_7_ZIP% e "%_ZLIB_SOURCE%" -y 1> nul && %_7_ZIP% x "%_ZLIB_SOURCE_FILENAME%.tar" -y 1> nul
 @del %_ZLIB_SOURCE_FILENAME%.tar
 
-%_CMAKE% -G %_COMPILER_NAME% -DCMAKE_INSTALL_PREFIX=%_GENERATED_PATH%/zlib %_GENERATED_PATH%/%_ZLIB_SOURCE_FILENAME%
+@if %_COMPILER_NAME% == "MinGW Makefiles" @(
+	@if not exist %_GENERATED_PATH%/release (
+		@mkdir "%_GENERATED_PATH%/release"
+	)
+	@cd release
+	
+	%_CMAKE% -G %_COMPILER_NAME% -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=%_GENERATED_PATH%/zlib %_GENERATED_PATH%/%_ZLIB_SOURCE_FILENAME%
 
-%_CMAKE% --build . --target install --config Release
+	%_CMAKE% --build . --target install
+	
+	@cd ..
+) else (
+	%_CMAKE% -G %_COMPILER_NAME% -DCMAKE_INSTALL_PREFIX=%_GENERATED_PATH%/zlib %_GENERATED_PATH%/%_ZLIB_SOURCE_FILENAME%
+
+	%_CMAKE% --build . --target install --config Release
+)
 
 @if "%1" neq "-nopause" (
 	@pause
